WIP: Linux port  - Fixing Mono initialization
Former-commit-id: 0cd8d0e66ca12a2794227c4871c781be7960eba4

+++ Source/BansheeCore/RenderAPI/BsGpuPipelineParamInfo.cpp
@@ -211,7 +211,7 @@ namespace bs
 			// Allow sampler states & textures/buffers to share the same slot, as some APIs combine them
 			if(type == ParamType::SamplerState)
 			{
-				if (mSetInfos[set].slotSamplers[slot] != -1)
+				if (mSetInfos[set].slotSamplers[slot] != (UINT32)-1)
 					return mSetInfos[set].slotSamplers[slot];
 			}
